pyproject.toml
[tool.poetry]
name = "artificialfriend"
version = "0.1.0"
description = ""
authors = ["sweemeng <sweester@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.190"
python-dotenv = "^1.0.0"
openai = "^0.27.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

